Struct google_androidenterprise1::api::AndroidEnterprise

pub struct AndroidEnterprise<C> { /* fields omitted */ }

Central instance to access all AndroidEnterprise related resource activities

Examples

Instantiate a new hub

extern crate hyper;
extern crate hyper_rustls;
extern crate yup_oauth2 as oauth2;
extern crate google_androidenterprise1 as androidenterprise1;
use androidenterprise1::{Result, Error};
use std::default::Default;
use oauth2;
use androidenterprise1::AndroidEnterprise;
 
// Get an ApplicationSecret instance by some means. It contains the `client_id` and 
// `client_secret`, among other things.
let secret: oauth2::ApplicationSecret = Default::default();
// Instantiate the authenticator. It will choose a suitable authentication flow for you, 
// unless you replace  `None` with the desired Flow.
// Provide your own `AuthenticatorDelegate` to adjust the way it operates and get feedback about 
// what's going on. You probably want to bring in your own `TokenStorage` to persist tokens and
// retrieve them from storage.
let auth = yup_oauth2::InstalledFlowAuthenticator::builder(
        secret,
        yup_oauth2::InstalledFlowReturnMethod::HTTPRedirect,
    ).build().await.unwrap();
let mut hub = AndroidEnterprise::new(hyper::Client::builder().build(hyper_rustls::HttpsConnector::with_native_roots()), auth);
// You can configure optional parameters by calling the respective setters at will, and
// execute the final call using `doit()`.
// Values shown here are possibly random and not representative !
let result = hub.enterprises().complete_signup()
             .enterprise_token("voluptua.")
             .completion_token("At")
             .doit().await;
 
match result {
    Err(e) => match e {
        // The Error enum provides details about what exactly happened.
        // You can also just use its `Debug`, `Display` or `Error` traits
         Error::HttpError(_)
        |Error::Io(_)
        |Error::MissingAPIKey
        |Error::MissingToken(_)
        |Error::Cancelled
        |Error::UploadSizeLimitExceeded(_, _)
        |Error::Failure(_)
        |Error::BadRequest(_)
        |Error::FieldClash(_)
        |Error::JsonDecodeError(_, _) => println!("{}", e),
    },
    Ok(res) => println!("Success: {:?}", res),
}

Implementations

impl<'a, C> AndroidEnterprise<C> where
C: BorrowMut<Client<HttpsConnector<HttpConnector>, Body>>,

pub fn new(
client: C,
authenticator: Authenticator<HttpsConnector<HttpConnector>>
) -> AndroidEnterprise<C>

pub fn devices(&'a self) -> DeviceMethods<'a, C>

pub fn enterprises(&'a self) -> EnterpriseMethods<'a, C>

pub fn entitlements(&'a self) -> EntitlementMethods<'a, C>

pub fn grouplicenses(&'a self) -> GrouplicenseMethods<'a, C>

pub fn grouplicenseusers(&'a self) -> GrouplicenseuserMethods<'a, C>

pub fn installs(&'a self) -> InstallMethods<'a, C>

pub fn managedconfigurationsfordevice(
&'a self
) -> ManagedconfigurationsfordeviceMethods<'a, C>

pub fn managedconfigurationsforuser(
&'a self
) -> ManagedconfigurationsforuserMethods<'a, C>

pub fn managedconfigurationssettings(
&'a self
) -> ManagedconfigurationssettingMethods<'a, C>

pub fn permissions(&'a self) -> PermissionMethods<'a, C>

pub fn products(&'a self) -> ProductMethods<'a, C>

pub fn serviceaccountkeys(&'a self) -> ServiceaccountkeyMethods<'a, C>

pub fn storelayoutclusters(&'a self) -> StorelayoutclusterMethods<'a, C>

pub fn storelayoutpages(&'a self) -> StorelayoutpageMethods<'a, C>

pub fn users(&'a self) -> UserMethods<'a, C>

pub fn webapps(&'a self) -> WebappMethods<'a, C>

pub fn user_agent(&mut self, agent_name: String) -> String

Set the user-agent header field to use in all requests to the server. It defaults to google-api-rust-client/2.0.0.

Returns the previously set user-agent.

pub fn base_url(&mut self, new_base_url: String) -> String

Set the base url to use in all requests to the server. It defaults to https://androidenterprise.googleapis.com/.

Returns the previously set base url.

pub fn root_url(&mut self, new_root_url: String) -> String

Set the root url to use in all requests to the server. It defaults to https://androidenterprise.googleapis.com/.

Returns the previously set root url.