Struct google_androidenterprise1::api::EnterpriseMethods

pub struct EnterpriseMethods<'a, C> where
    C: 'a,  { /* fields omitted */ }

A builder providing access to all methods supported on enterprise resources. It is not used directly, but through the AndroidEnterprise hub.

Example

Instantiate a resource builder

extern crate hyper;
extern crate hyper_rustls;
extern crate yup_oauth2 as oauth2;
extern crate google_androidenterprise1 as androidenterprise1;
 
use std::default::Default;
use oauth2;
use androidenterprise1::AndroidEnterprise;
 
let secret: oauth2::ApplicationSecret = Default::default();
let auth = yup_oauth2::InstalledFlowAuthenticator::builder(
        secret,
        yup_oauth2::InstalledFlowReturnMethod::HTTPRedirect,
    ).build().await.unwrap();
let mut hub = AndroidEnterprise::new(hyper::Client::builder().build(hyper_rustls::HttpsConnector::with_native_roots()), auth);
// Usually you wouldn't bind this to a variable, but keep calling *CallBuilders*
// like `acknowledge_notification_set(...)`, `complete_signup(...)`, `create_web_token(...)`, `enroll(...)`, `generate_signup_url(...)`, `get(...)`, `get_service_account(...)`, `get_store_layout(...)`, `list(...)`, `pull_notification_set(...)`, `send_test_push_notification(...)`, `set_account(...)`, `set_store_layout(...)` and `unenroll(...)`
// to build up your call.
let rb = hub.enterprises();

Implementations

impl<'a, C> EnterpriseMethods<'a, C>

pub fn acknowledge_notification_set(
&self
) -> EnterpriseAcknowledgeNotificationSetCall<'a, C>

Create a builder to help you perform the following task:

Acknowledges notifications that were received from Enterprises.PullNotificationSet to prevent subsequent calls from returning the same notifications.

pub fn complete_signup(&self) -> EnterpriseCompleteSignupCall<'a, C>

Create a builder to help you perform the following task:

Completes the signup flow, by specifying the Completion token and Enterprise token. This request must not be called multiple times for a given Enterprise Token.

pub fn create_web_token(
&self,
request: AdministratorWebTokenSpec,
enterprise_id: &str
) -> EnterpriseCreateWebTokenCall<'a, C>

Create a builder to help you perform the following task:

Returns a unique token to access an embeddable UI. To generate a web UI, pass the generated token into the managed Google Play javascript API. Each token may only be used to start one UI session. See the javascript API documentation for further information.

Arguments

request - No description provided.

enterpriseId - The ID of the enterprise.

pub fn enroll(
&self,
request: Enterprise,
token: &str
) -> EnterpriseEnrollCall<'a, C>

Create a builder to help you perform the following task:

Enrolls an enterprise with the calling EMM.

Arguments

request - No description provided.

token - Required. The token provided by the enterprise to register the EMM.

pub fn generate_signup_url(&self) -> EnterpriseGenerateSignupUrlCall<'a, C>

Create a builder to help you perform the following task:

Generates a sign-up URL.

pub fn get(&self, enterprise_id: &str) -> EnterpriseGetCall<'a, C>

Create a builder to help you perform the following task:

Retrieves the name and domain of an enterprise.

Arguments

enterpriseId - The ID of the enterprise.

pub fn get_service_account(
&self,
enterprise_id: &str
) -> EnterpriseGetServiceAccountCall<'a, C>

Create a builder to help you perform the following task:

Returns a service account and credentials. The service account can be bound to the enterprise by calling setAccount. The service account is unique to this enterprise and EMM, and will be deleted if the enterprise is unbound. The credentials contain private key data and are not stored server-side. This method can only be called after calling Enterprises.Enroll or Enterprises.CompleteSignup, and before Enterprises.SetAccount; at other times it will return an error. Subsequent calls after the first will generate a new, unique set of credentials, and invalidate the previously generated credentials. Once the service account is bound to the enterprise, it can be managed using the serviceAccountKeys resource.

Arguments

enterpriseId - The ID of the enterprise.

pub fn get_store_layout(
&self,
enterprise_id: &str
) -> EnterpriseGetStoreLayoutCall<'a, C>

Create a builder to help you perform the following task:

Returns the store layout for the enterprise. If the store layout has not been set, returns "basic" as the store layout type and no homepage.

Arguments

enterpriseId - The ID of the enterprise.

pub fn list(&self, domain: &str) -> EnterpriseListCall<'a, C>

Create a builder to help you perform the following task:

Looks up an enterprise by domain name. This is only supported for enterprises created via the Google-initiated creation flow. Lookup of the id is not needed for enterprises created via the EMM-initiated flow since the EMM learns the enterprise ID in the callback specified in the Enterprises.generateSignupUrl call.

Arguments

domain - Required. The exact primary domain name of the enterprise to look up.

pub fn pull_notification_set(&self) -> EnterprisePullNotificationSetCall<'a, C>

Create a builder to help you perform the following task:

Pulls and returns a notification set for the enterprises associated with the service account authenticated for the request. The notification set may be empty if no notification are pending. A notification set returned needs to be acknowledged within 20 seconds by calling Enterprises.AcknowledgeNotificationSet, unless the notification set is empty. Notifications that are not acknowledged within the 20 seconds will eventually be included again in the response to another PullNotificationSet request, and those that are never acknowledged will ultimately be deleted according to the Google Cloud Platform Pub/Sub system policy. Multiple requests might be performed concurrently to retrieve notifications, in which case the pending notifications (if any) will be split among each caller, if any are pending. If no notifications are present, an empty notification list is returned. Subsequent requests may return more notifications once they become available.

pub fn send_test_push_notification(
&self,
enterprise_id: &str
) -> EnterpriseSendTestPushNotificationCall<'a, C>

Create a builder to help you perform the following task:

Sends a test notification to validate the EMM integration with the Google Cloud Pub/Sub service for this enterprise.

Arguments

enterpriseId - The ID of the enterprise.

pub fn set_account(
&self,
request: EnterpriseAccount,
enterprise_id: &str
) -> EnterpriseSetAccountCall<'a, C>

Create a builder to help you perform the following task:

Sets the account that will be used to authenticate to the API as the enterprise.

Arguments

request - No description provided.

enterpriseId - The ID of the enterprise.

pub fn set_store_layout(
&self,
request: StoreLayout,
enterprise_id: &str
) -> EnterpriseSetStoreLayoutCall<'a, C>

Create a builder to help you perform the following task:

Sets the store layout for the enterprise. By default, storeLayoutType is set to "basic" and the basic store layout is enabled. The basic layout only contains apps approved by the admin, and that have been added to the available product set for a user (using the setAvailableProductSet call). Apps on the page are sorted in order of their product ID value. If you create a custom store layout (by setting storeLayoutType = "custom" and setting a homepage), the basic store layout is disabled.

Arguments

request - No description provided.

enterpriseId - The ID of the enterprise.

pub fn unenroll(&self, enterprise_id: &str) -> EnterpriseUnenrollCall<'a, C>

Create a builder to help you perform the following task:

Unenrolls an enterprise from the calling EMM.

Arguments

enterpriseId - The ID of the enterprise.