Struct google-androidenterprise1::AndroidEnterprise

pub struct AndroidEnterprise<C, NC, A> {
    // some fields omitted
}

Central instance to access all AndroidEnterprise related resource activities

Examples

Instantiate a new hub

extern crate hyper;
extern crate "yup-oauth2" as oauth2;
extern crate serde;
extern crate "google-androidenterprise1" as androidenterprise1;
use androidenterprise1::Enterprise;
use androidenterprise1::Result;
use std::default::Default;
use oauth2::{Authenticator, DefaultAuthenticatorDelegate, ApplicationSecret, MemoryStorage};
use androidenterprise1::AndroidEnterprise;
 
// Get an ApplicationSecret instance by some means. It contains the `client_id` and 
// `client_secret`, among other things.
let secret: ApplicationSecret = Default::default();
// Instantiate the authenticator. It will choose a suitable authentication flow for you, 
// unless you replace  `None` with the desired Flow.
// Provide your own `AuthenticatorDelegate` to adjust the way it operates and get feedback about 
// what's going on. You probably want to bring in your own `TokenStorage` to persist tokens and
// retrieve them from storage.
let auth = Authenticator::new(&secret, DefaultAuthenticatorDelegate,
                              hyper::Client::new(),
                              <MemoryStorage as Default>::default(), None);
let mut hub = AndroidEnterprise::new(hyper::Client::new(), auth);
// As the method needs a request, you would usually fill it with the desired information
// into the respective structure. Some of the parts shown here might not be applicable !
// Values shown here are possibly random and not representative !
let mut req: Enterprise = Default::default();
 
// You can configure optional parameters by calling the respective setters at will, and
// execute the final call using `doit()`.
// Values shown here are possibly random and not representative !
let result = hub.enterprises().insert(&req, "token")
             .doit();
 
match result {
    Result::HttpError(err) => println!("HTTPERROR: {:?}", err),
    Result::MissingAPIKey => println!("Missing API Key"),
    Result::MissingToken => println!("Missing Token"),
    Result::Failure(_) => println!("General Failure (Response doesn't print)"),
    Result::FieldClash(clashed_field) => println!("FIELD CLASH: {:?}", clashed_field),
    Result::JsonDecodeError(err) => println!("Json failed to decode: {:?}", err),
    Result::Success(_) => println!("Success (value doesn't print)"),
}

Methods

impl<'a, C, NC, A> AndroidEnterprise<C, NC, A> where NC: NetworkConnector, C: BorrowMut<Client<NC>>, A: GetToken

fn new(client: C, authenticator: A) -> AndroidEnterprise<C, NC, A>

fn collections(&'a self) -> CollectionMethodsBuilder<'a, C, NC, A>

fn collectionviewers(&'a self) -> CollectionviewerMethodsBuilder<'a, C, NC, A>

fn devices(&'a self) -> DeviceMethodsBuilder<'a, C, NC, A>

fn enterprises(&'a self) -> EnterpriseMethodsBuilder<'a, C, NC, A>

fn entitlements(&'a self) -> EntitlementMethodsBuilder<'a, C, NC, A>

fn grouplicenses(&'a self) -> GrouplicenseMethodsBuilder<'a, C, NC, A>

fn grouplicenseusers(&'a self) -> GrouplicenseuserMethodsBuilder<'a, C, NC, A>

fn installs(&'a self) -> InstallMethodsBuilder<'a, C, NC, A>

fn permissions(&'a self) -> PermissionMethodsBuilder<'a, C, NC, A>

fn products(&'a self) -> ProductMethodsBuilder<'a, C, NC, A>

fn users(&'a self) -> UserMethodsBuilder<'a, C, NC, A>

fn user_agent(&mut self, agent_name: String) -> String

Set the user-agent header field to use in all requests to the server. It defaults to google-api-rust-client/0.0.1.

Returns the previously set user-agent.