Struct google_androidenterprise1::AppRestrictionsSchemaRestriction

pub struct AppRestrictionsSchemaRestriction {
    pub description: Option<String>,
    pub title: Option<String>,
    pub default_value: Option<AppRestrictionsSchemaRestrictionRestrictionValue>,
    pub nested_restriction: Option<Vec<AppRestrictionsSchemaRestriction>>,
    pub entry_value: Option<Vec<String>>,
    pub key: Option<String>,
    pub entry: Option<Vec<String>>,
    pub restriction_type: Option<String>,
}

A restriction in the App Restriction Schema represents a piece of configuration that may be pre-applied.

This type is not used in any activity, and only used as part of another schema.

Fields
description: Option<String>

A longer description of the restriction, giving more detail of what it affects.

title: Option<String>

The name of the restriction.

default_value: Option<AppRestrictionsSchemaRestrictionRestrictionValue>

The default value of the restriction.

nested_restriction: Option<Vec<AppRestrictionsSchemaRestriction>>

For bundle or bundleArray restrictions, the list of nested restrictions.

entry_value: Option<Vec<String>>

For choice or multiselect restrictions, the list of possible entries' machine-readable values.

key: Option<String>

The unique key that the product uses to identify the restriction, e.g. "com.google.android.gm.fieldname".

entry: Option<Vec<String>>

For choice or multiselect restrictions, the list of possible entries' human-readable names.

restriction_type: Option<String>

The type of the restriction.