Struct google_androidenterprise1::AppRestrictionsSchemaRestrictionRestrictionValue

pub struct AppRestrictionsSchemaRestrictionRestrictionValue {
    pub value_bool: Option<bool>,
    pub value_integer: Option<i32>,
    pub type_: Option<String>,
    pub value_multiselect: Option<Vec<String>>,
    pub value_string: Option<String>,
}

A typed value for the restriction.

This type is not used in any activity, and only used as part of another schema.

Fields

value_bool

The boolean value - this will only be present if type is bool.

value_integer

The integer value - this will only be present if type is integer.

type_

The type of the value being provided.

value_multiselect

The list of string values - this will only be present if type is multiselect.

value_string

The string value - this will be present for types string, choice and hidden.