Struct google_androidenterprise1::CollectionviewerMethods

pub struct CollectionviewerMethods<'a, C, A> where C: 'a, A: 'a {
    // some fields omitted
}

A builder providing access to all methods supported on collectionviewer resources. It is not used directly, but through the AndroidEnterprise hub.

Example

Instantiate a resource builder

extern crate hyper;
extern crate yup_oauth2 as oauth2;
extern crate google_androidenterprise1 as androidenterprise1;
 
use std::default::Default;
use oauth2::{Authenticator, DefaultAuthenticatorDelegate, ApplicationSecret, MemoryStorage};
use androidenterprise1::AndroidEnterprise;
 
let secret: ApplicationSecret = Default::default();
let auth = Authenticator::new(&secret, DefaultAuthenticatorDelegate,
                              hyper::Client::new(),
                              <MemoryStorage as Default>::default(), None);
let mut hub = AndroidEnterprise::new(hyper::Client::new(), auth);
// Usually you wouldn't bind this to a variable, but keep calling *CallBuilders*
// like `delete(...)`, `get(...)`, `list(...)`, `patch(...)` and `update(...)`
// to build up your call.
let rb = hub.collectionviewers();

Methods

impl<'a, C, A> CollectionviewerMethods<'a, C, A>

fn delete(&self, enterprise_id: &str, collection_id: &str, user_id: &str) -> CollectionviewerDeleteCall<'a, C, A>

Create a builder to help you perform the following task:

Removes the user from the list of those specifically allowed to see the collection. If the collection's visibility is set to viewersOnly then only such users will see the collection.

Arguments

enterpriseId - The ID of the enterprise.

collectionId - The ID of the collection.

userId - The ID of the user.

fn patch(&self, request: User, enterprise_id: &str, collection_id: &str, user_id: &str) -> CollectionviewerPatchCall<'a, C, A>

Create a builder to help you perform the following task:

Adds the user to the list of those specifically allowed to see the collection. If the collection's visibility is set to viewersOnly then only such users will see the collection. This method supports patch semantics.

Arguments

request - No description provided.

enterpriseId - The ID of the enterprise.

collectionId - The ID of the collection.

userId - The ID of the user.

fn get(&self, enterprise_id: &str, collection_id: &str, user_id: &str) -> CollectionviewerGetCall<'a, C, A>

Create a builder to help you perform the following task:

Retrieves the ID of the user if they have been specifically allowed to see the collection. If the collection's visibility is set to viewersOnly then only these users will see the collection.

Arguments

enterpriseId - The ID of the enterprise.

collectionId - The ID of the collection.

userId - The ID of the user.

fn list(&self, enterprise_id: &str, collection_id: &str) -> CollectionviewerListCall<'a, C, A>

Create a builder to help you perform the following task:

Retrieves the IDs of the users who have been specifically allowed to see the collection. If the collection's visibility is set to viewersOnly then only these users will see the collection.

Arguments

enterpriseId - The ID of the enterprise.

collectionId - The ID of the collection.

fn update(&self, request: User, enterprise_id: &str, collection_id: &str, user_id: &str) -> CollectionviewerUpdateCall<'a, C, A>

Create a builder to help you perform the following task:

Adds the user to the list of those specifically allowed to see the collection. If the collection's visibility is set to viewersOnly then only such users will see the collection.

Arguments

request - No description provided.

enterpriseId - The ID of the enterprise.

collectionId - The ID of the collection.

userId - The ID of the user.