Struct google_androidenterprise1::EnterpriseMethods

pub struct EnterpriseMethods<'a, C, A> where C: 'a, A: 'a {
    // some fields omitted
}

A builder providing access to all methods supported on enterprise resources. It is not used directly, but through the AndroidEnterprise hub.

Example

Instantiate a resource builder

extern crate hyper;
extern crate yup_oauth2 as oauth2;
extern crate google_androidenterprise1 as androidenterprise1;
 
use std::default::Default;
use oauth2::{Authenticator, DefaultAuthenticatorDelegate, ApplicationSecret, MemoryStorage};
use androidenterprise1::AndroidEnterprise;
 
let secret: ApplicationSecret = Default::default();
let auth = Authenticator::new(&secret, DefaultAuthenticatorDelegate,
                              hyper::Client::new(),
                              <MemoryStorage as Default>::default(), None);
let mut hub = AndroidEnterprise::new(hyper::Client::new(), auth);
// Usually you wouldn't bind this to a variable, but keep calling *CallBuilders*
// like `acknowledge_notification_set(...)`, `complete_signup(...)`, `delete(...)`, `enroll(...)`, `generate_signup_url(...)`, `get(...)`, `get_service_account(...)`, `get_store_layout(...)`, `insert(...)`, `list(...)`, `pull_notification_set(...)`, `send_test_push_notification(...)`, `set_account(...)`, `set_store_layout(...)` and `unenroll(...)`
// to build up your call.
let rb = hub.enterprises();

Methods

impl<'a, C, A> EnterpriseMethods<'a, C, A>

fn insert(&self, request: Enterprise, token: &str) -> EnterpriseInsertCall<'a, C, A>

Create a builder to help you perform the following task:

Establishes the binding between the EMM and an enterprise. This is now deprecated; use enroll instead.

Arguments

request - No description provided.

token - The token provided by the enterprise to register the EMM.

fn set_store_layout(&self, request: StoreLayout, enterprise_id: &str) -> EnterpriseSetStoreLayoutCall<'a, C, A>

Create a builder to help you perform the following task:

Sets the store layout for the enterprise.

Arguments

request - No description provided.

enterpriseId - The ID of the enterprise.

fn pull_notification_set(&self) -> EnterprisePullNotificationSetCall<'a, C, A>

Create a builder to help you perform the following task:

Pulls and returns a notification set for the enterprises associated with the service account authenticated for the request. The notification set may be empty if no notification are pending. A notification set returned needs to be acknowledged within 20 seconds by calling Enterprises.AcknowledgeNotificationSet, unless the notification set is empty. Notifications that are not acknowledged within the 20 seconds will eventually be included again in the response to another PullNotificationSet request, and those that are never acknowledged will ultimately be deleted according to the Google Cloud Platform Pub/Sub system policy. Multiple requests might be performed concurrently to retrieve notifications, in which case the pending notifications (if any) will be split among each caller, if any are pending.

fn enroll(&self, request: Enterprise, token: &str) -> EnterpriseEnrollCall<'a, C, A>

Create a builder to help you perform the following task:

Enrolls an enterprise with the calling EMM.

Arguments

request - No description provided.

token - The token provided by the enterprise to register the EMM.

fn set_account(&self, request: EnterpriseAccount, enterprise_id: &str) -> EnterpriseSetAccountCall<'a, C, A>

Create a builder to help you perform the following task:

Set the account that will be used to authenticate to the API as the enterprise.

Arguments

request - No description provided.

enterpriseId - The ID of the enterprise.

fn list(&self, domain: &str) -> EnterpriseListCall<'a, C, A>

Create a builder to help you perform the following task:

Looks up an enterprise by domain name. This is only supported for enterprises created via the Google-initiated creation flow. Lookup of the id is not needed for enterprises created via the EMM-initiated flow since the EMM learns the enterprise ID in the callback specified in the Enterprises.generateSignupUrl call.

Arguments

domain - The exact primary domain name of the enterprise to look up.

fn get_store_layout(&self, enterprise_id: &str) -> EnterpriseGetStoreLayoutCall<'a, C, A>

Create a builder to help you perform the following task:

Returns the store layout for the enterprise. If the store layout has not been set, or if the store layout has no homepageId set, returns a NOT_FOUND error.

Arguments

enterpriseId - The ID of the enterprise.

fn complete_signup(&self) -> EnterpriseCompleteSignupCall<'a, C, A>

Create a builder to help you perform the following task:

Completes the signup flow, by specifying the Completion token and Enterprise token. This request must not be called multiple times for a given Enterprise Token.

fn get_service_account(&self, enterprise_id: &str) -> EnterpriseGetServiceAccountCall<'a, C, A>

Create a builder to help you perform the following task:

Returns a service account and credentials. The service account can be bound to the enterprise by calling setAccount. The service account is unique to this enterprise and EMM, and will be deleted if the enterprise is unbound. The credentials contain private key data and are not stored server-side.

This method can only be called after calling Enterprises.Enroll or Enterprises.CompleteSignup, and before Enterprises.SetAccount; at other times it will return an error.

Subsequent calls after the first will generate a new, unique set of credentials, and invalidate the previously generated credentials.

Once the service account is bound to the enterprise, it can be managed using the serviceAccountKeys resource.

Arguments

enterpriseId - The ID of the enterprise.

fn acknowledge_notification_set(&self) -> EnterpriseAcknowledgeNotificationSetCall<'a, C, A>

Create a builder to help you perform the following task:

Acknowledges notifications that were received from Enterprises.PullNotificationSet to prevent subsequent calls from returning the same notifications.

fn unenroll(&self, enterprise_id: &str) -> EnterpriseUnenrollCall<'a, C, A>

Create a builder to help you perform the following task:

Unenrolls an enterprise from the calling EMM.

Arguments

enterpriseId - The ID of the enterprise.

fn send_test_push_notification(&self, enterprise_id: &str) -> EnterpriseSendTestPushNotificationCall<'a, C, A>

Create a builder to help you perform the following task:

Sends a test push notification to validate the EMM integration with the Google Cloud Pub/Sub service for this enterprise.

Arguments

enterpriseId - The ID of the enterprise.

fn get(&self, enterprise_id: &str) -> EnterpriseGetCall<'a, C, A>

Create a builder to help you perform the following task:

Retrieves the name and domain of an enterprise.

Arguments

enterpriseId - The ID of the enterprise.

fn delete(&self, enterprise_id: &str) -> EnterpriseDeleteCall<'a, C, A>

Create a builder to help you perform the following task:

Deletes the binding between the EMM and enterprise. This is now deprecated; use this to unenroll customers that were previously enrolled with the 'insert' call, then enroll them again with the 'enroll' call.

Arguments

enterpriseId - The ID of the enterprise.

fn generate_signup_url(&self) -> EnterpriseGenerateSignupUrlCall<'a, C, A>

Create a builder to help you perform the following task:

Generates a sign-up URL.