Struct google_androidenterprise1::GroupLicense

pub struct GroupLicense {
    pub kind: Option<String>,
    pub num_purchased: Option<i32>,
    pub num_provisioned: Option<i32>,
    pub acquisition_kind: Option<String>,
    pub approval: Option<String>,
    pub product_id: Option<String>,
}

A group license object indicates a product that an enterprise admin has approved for use in the enterprise. The product may be free or paid. For free products, a group license object is created in these cases: if the enterprise admin approves a product in Google Play, if the product is added to a collection, or if an entitlement for the product is created for a user via the API. For paid products, a group license object is only created as part of the first bulk purchase of that product in Google Play by the enterprise admin.

The API can be used to query group licenses; the available information includes the total number of licenses purchased (for paid products) and the total number of licenses that have been provisioned, that is, the total number of user entitlements in existence for the product.

Group license objects are never deleted. If, for example, a free app is added to a collection and then removed, the group license will remain, allowing the enterprise admin to keep track of any remaining entitlements. An enterprise admin may indicate they are no longer interested in the group license by marking it as unapproved in Google Play.

Activities

This type is used in activities, which are methods you may call on this type or where this type is involved in. The list links the activity name, along with information about where it is used (one of request and response).

get grouplicenses (response)

Fields

kind

Identifies what kind of resource this is. Value: the fixed string "androidenterprise#groupLicense".

num_purchased

The number of purchased licenses (possibly in multiple purchases). If this field is omitted then there is no limit on the number of licenses that can be provisioned (e.g. if the acquisition kind is "free").

num_provisioned

The total number of provisioned licenses for this product. Returned by read operations, but ignored in write operations.

acquisition_kind

How this group license was acquired. "bulkPurchase" means that this group license object was created because the enterprise purchased licenses for this product; this is "free" otherwise (for free products).

approval

Whether the product to which this group license relates is currently approved by the enterprise, as either "approved" or "unapproved". Products are approved when a group license is first created, but this approval may be revoked by an enterprise admin via Google Play. Unapproved products will not be visible to end users in collections and new entitlements to them should not normally be created.

product_id

The ID of the product that the license is for, e.g. "app:com.google.android.gm".