Struct google_androidenterprise1::GrouplicenseuserMethods

pub struct GrouplicenseuserMethods<'a, C, A> where C: 'a, A: 'a {
    // some fields omitted
}

A builder providing access to all methods supported on grouplicenseuser resources. It is not used directly, but through the AndroidEnterprise hub.

Example

Instantiate a resource builder

extern crate hyper;
extern crate yup_oauth2 as oauth2;
extern crate google_androidenterprise1 as androidenterprise1;
 
use std::default::Default;
use oauth2::{Authenticator, DefaultAuthenticatorDelegate, ApplicationSecret, MemoryStorage};
use androidenterprise1::AndroidEnterprise;
 
let secret: ApplicationSecret = Default::default();
let auth = Authenticator::new(&secret, DefaultAuthenticatorDelegate,
                              hyper::Client::new(),
                              <MemoryStorage as Default>::default(), None);
let mut hub = AndroidEnterprise::new(hyper::Client::new(), auth);
// Usually you wouldn't bind this to a variable, but keep calling *CallBuilders*
// like `list(...)`
// to build up your call.
let rb = hub.grouplicenseusers();

Methods

impl<'a, C, A> GrouplicenseuserMethods<'a, C, A>

fn list(&self, enterprise_id: &str, group_license_id: &str) -> GrouplicenseuserListCall<'a, C, A>

Create a builder to help you perform the following task:

Retrieves the IDs of the users who have been granted entitlements under the license.

Arguments

enterpriseId - The ID of the enterprise.

groupLicenseId - The ID of the product the group license is for, e.g. "app:com.google.android.gm".