Struct google_androidenterprise1::Product

pub struct Product {
    pub icon_url: Option<String>,
    pub kind: Option<String>,
    pub title: Option<String>,
    pub details_url: Option<String>,
    pub work_details_url: Option<String>,
    pub author_name: Option<String>,
    pub product_id: Option<String>,
}

A product represents an app in the Google Play Store that is available to at least some users in the enterprise. (Some apps are restricted to a single enterprise, and no information about them is made available outside that enterprise.)

The information provided for each product (localized name, icon, link to the full Google Play details page) is intended to allow a basic representation of the product within an MDM user interface.

Activities

This type is used in activities, which are methods you may call on this type or where this type is involved in. The list links the activity name, along with information about where it is used (one of request and response).

get products (response)

update permissions products (none)

get app restrictions schema products (none)

get permissions products (none)

Fields

icon_url

A link to an image that can be used as an icon for the product.

kind

Identifies what kind of resource this is. Value: the fixed string "androidenterprise#product".

title

The name of the product.

details_url

A link to the (consumer) Google Play details page for the product.

work_details_url

A link to the Google Play for Work details page for the product, for use by an Enterprise administrator.

author_name

The name of the author of the product (e.g. the app developer).

product_id

A string of the form "app: " - e.g. "app:com.google.android.gm" represents the GMail app.