Struct google_androidenterprise1::StorelayoutpageMethods

pub struct StorelayoutpageMethods<'a, C, A> where C: 'a, A: 'a {
    // some fields omitted
}

A builder providing access to all methods supported on storelayoutpage resources. It is not used directly, but through the AndroidEnterprise hub.

Example

Instantiate a resource builder

extern crate hyper;
extern crate yup_oauth2 as oauth2;
extern crate google_androidenterprise1 as androidenterprise1;
 
use std::default::Default;
use oauth2::{Authenticator, DefaultAuthenticatorDelegate, ApplicationSecret, MemoryStorage};
use androidenterprise1::AndroidEnterprise;
 
let secret: ApplicationSecret = Default::default();
let auth = Authenticator::new(&secret, DefaultAuthenticatorDelegate,
                              hyper::Client::new(),
                              <MemoryStorage as Default>::default(), None);
let mut hub = AndroidEnterprise::new(hyper::Client::new(), auth);
// Usually you wouldn't bind this to a variable, but keep calling *CallBuilders*
// like `delete(...)`, `get(...)`, `insert(...)`, `list(...)`, `patch(...)` and `update(...)`
// to build up your call.
let rb = hub.storelayoutpages();

Methods

impl<'a, C, A> StorelayoutpageMethods<'a, C, A>

fn delete(&self, enterprise_id: &str, page_id: &str) -> StorelayoutpageDeleteCall<'a, C, A>

Create a builder to help you perform the following task:

Deletes a store page.

Arguments

enterpriseId - The ID of the enterprise.

pageId - The ID of the page.

fn patch(&self, request: StorePage, enterprise_id: &str, page_id: &str) -> StorelayoutpagePatchCall<'a, C, A>

Create a builder to help you perform the following task:

Updates the content of a store page. This method supports patch semantics.

Arguments

request - No description provided.

enterpriseId - The ID of the enterprise.

pageId - The ID of the page.

fn insert(&self, request: StorePage, enterprise_id: &str) -> StorelayoutpageInsertCall<'a, C, A>

Create a builder to help you perform the following task:

Inserts a new store page.

Arguments

request - No description provided.

enterpriseId - The ID of the enterprise.

fn update(&self, request: StorePage, enterprise_id: &str, page_id: &str) -> StorelayoutpageUpdateCall<'a, C, A>

Create a builder to help you perform the following task:

Updates the content of a store page.

Arguments

request - No description provided.

enterpriseId - The ID of the enterprise.

pageId - The ID of the page.

fn get(&self, enterprise_id: &str, page_id: &str) -> StorelayoutpageGetCall<'a, C, A>

Create a builder to help you perform the following task:

Retrieves details of a store page.

Arguments

enterpriseId - The ID of the enterprise.

pageId - The ID of the page.

fn list(&self, enterprise_id: &str) -> StorelayoutpageListCall<'a, C, A>

Create a builder to help you perform the following task:

Retrieves the details of all pages in the store.

Arguments

enterpriseId - The ID of the enterprise.