Struct google_androidenterprise1::User

pub struct User {
    pub kind: Option<String>,
    pub display_name: Option<String>,
    pub primary_email: Option<String>,
    pub account_type: Option<String>,
    pub id: Option<String>,
    pub account_identifier: Option<String>,
    pub management_type: Option<String>,
}

A Users resource represents an account associated with an enterprise. The account may be specific to a device or to an individual user (who can then use the account across multiple devices). The account may provide access to Google Play for Work only, or to other Google services, depending on the identity model: - Google managed domain identity model requires synchronization to Google account sources (via primaryEmail). - Android for Work Accounts identity model provides a dynamic means for enterprises to create user or device accounts as needed. These accounts provide access to Google Play for Work only.

Activities

This type is used in activities, which are methods you may call on this type or where this type is involved in. The list links the activity name, along with information about where it is used (one of request and response).

revoke token users (none)

delete users (none)

update collectionviewers (request|response)

patch collectionviewers (request|response)

generate authentication token users (none)

list users (none)

update users (request|response)

get collectionviewers (response)

patch users (request|response)

set available product set users (none)

insert users (request|response)

generate token users (none)

get available product set users (none)

get users (response)

Fields
kind: Option<String>

Identifies what kind of resource this is. Value: the fixed string "androidenterprise#user".

display_name: Option<String>

The name that will appear in user interfaces. Setting this property is optional when creating EMM-managed users. If you do set this property, use something generic about the organization (such as "Example, Inc.") or your name (as EMM). Not used for Google-managed user accounts.

primary_email: Option<String>

The user's primary email address, for example, "jsmith@example.com". Will always be set for Google managed users and not set for EMM managed users.

account_type: Option<String>

The type of account that this user represents. A userAccount can be installed on multiple devices, but a deviceAccount is specific to a single device. An EMM-managed user (emmManaged) can be either type (userAccount, deviceAccount), but a Google-managed user (googleManaged) is always a userAccount.

id: Option<String>

The unique ID for the user.

account_identifier: Option<String>

A unique identifier you create for this user, such as "user342" or "asset#44418". Do not use personally identifiable information (PII) for this property. Must always be set for EMM-managed users. Not set for Google-managed users.

management_type: Option<String>

The entity that manages the user. With googleManaged users, the source of truth is Google so EMMs have to make sure a Google Account exists for the user. With emmManaged users, the EMM is in charge.