Struct google_androidenterprise1::UserMethods

pub struct UserMethods<'a, C, A> where
    C: 'a,
    A: 'a,  { /* fields omitted */ }

A builder providing access to all methods supported on user resources. It is not used directly, but through the AndroidEnterprise hub.

Example

Instantiate a resource builder

extern crate hyper;
extern crate hyper_rustls;
extern crate yup_oauth2 as oauth2;
extern crate google_androidenterprise1 as androidenterprise1;
 
use std::default::Default;
use oauth2::{Authenticator, DefaultAuthenticatorDelegate, ApplicationSecret, MemoryStorage};
use androidenterprise1::AndroidEnterprise;
 
let secret: ApplicationSecret = Default::default();
let auth = Authenticator::new(&secret, DefaultAuthenticatorDelegate,
                              hyper::Client::with_connector(hyper::net::HttpsConnector::new(hyper_rustls::TlsClient::new())),
                              <MemoryStorage as Default>::default(), None);
let mut hub = AndroidEnterprise::new(hyper::Client::with_connector(hyper::net::HttpsConnector::new(hyper_rustls::TlsClient::new())), auth);
// Usually you wouldn't bind this to a variable, but keep calling *CallBuilders*
// like `delete(...)`, `generate_authentication_token(...)`, `generate_token(...)`, `get(...)`, `get_available_product_set(...)`, `insert(...)`, `list(...)`, `patch(...)`, `revoke_device_access(...)`, `revoke_token(...)`, `set_available_product_set(...)` and `update(...)`
// to build up your call.
let rb = hub.users();

Methods

impl<'a, C, A> UserMethods<'a, C, A>

pub fn revoke_token(
&self,
enterprise_id: &str,
user_id: &str
) -> UserRevokeTokenCall<'a, C, A>

Create a builder to help you perform the following task:

Revokes a previously generated token (activation code) for the user.

Arguments

enterpriseId - The ID of the enterprise.

userId - The ID of the user.

pub fn update(
&self,
request: User,
enterprise_id: &str,
user_id: &str
) -> UserUpdateCall<'a, C, A>

Create a builder to help you perform the following task:

Updates the details of an EMM-managed user.

Can be used with EMM-managed users only (not Google managed users). Pass the new details in the Users resource in the request body. Only the displayName field can be changed. Other fields must either be unset or have the currently active value.

Arguments

request - No description provided.

enterpriseId - The ID of the enterprise.

userId - The ID of the user.

pub fn revoke_device_access(
&self,
enterprise_id: &str,
user_id: &str
) -> UserRevokeDeviceAccesCall<'a, C, A>

Create a builder to help you perform the following task:

Revokes access to all devices currently provisioned to the user. The user will no longer be able to use the managed Play store on any of their managed devices.

This call only works with EMM-managed accounts.

Arguments

enterpriseId - The ID of the enterprise.

userId - The ID of the user.

pub fn patch(
&self,
request: User,
enterprise_id: &str,
user_id: &str
) -> UserPatchCall<'a, C, A>

Create a builder to help you perform the following task:

Updates the details of an EMM-managed user.

Can be used with EMM-managed users only (not Google managed users). Pass the new details in the Users resource in the request body. Only the displayName field can be changed. Other fields must either be unset or have the currently active value. This method supports patch semantics.

Arguments

request - No description provided.

enterpriseId - The ID of the enterprise.

userId - The ID of the user.

pub fn set_available_product_set(
&self,
request: ProductSet,
enterprise_id: &str,
user_id: &str
) -> UserSetAvailableProductSetCall<'a, C, A>

Create a builder to help you perform the following task:

Modifies the set of products that a user is entitled to access (referred to as whitelisted products). Only products that are approved or products that were previously approved (products with revoked approval) can be whitelisted.

Arguments

request - No description provided.

enterpriseId - The ID of the enterprise.

userId - The ID of the user.

pub fn get_available_product_set(
&self,
enterprise_id: &str,
user_id: &str
) -> UserGetAvailableProductSetCall<'a, C, A>

Create a builder to help you perform the following task:

Retrieves the set of products a user is entitled to access.

Arguments

enterpriseId - The ID of the enterprise.

userId - The ID of the user.

pub fn get(&self, enterprise_id: &str, user_id: &str) -> UserGetCall<'a, C, A>

Create a builder to help you perform the following task:

Retrieves a user's details.

Arguments

enterpriseId - The ID of the enterprise.

userId - The ID of the user.

pub fn insert(
&self,
request: User,
enterprise_id: &str
) -> UserInsertCall<'a, C, A>

Create a builder to help you perform the following task:

Creates a new EMM-managed user.

The Users resource passed in the body of the request should include an accountIdentifier and an accountType. If a corresponding user already exists with the same account identifier, the user will be updated with the resource. In this case only the displayName field can be changed.

Arguments

request - No description provided.

enterpriseId - The ID of the enterprise.

pub fn generate_token(
&self,
enterprise_id: &str,
user_id: &str
) -> UserGenerateTokenCall<'a, C, A>

Create a builder to help you perform the following task:

Generates a token (activation code) to allow this user to configure their managed account in the Android Setup Wizard. Revokes any previously generated token.

This call only works with Google managed accounts.

Arguments

enterpriseId - The ID of the enterprise.

userId - The ID of the user.

pub fn delete(
&self,
enterprise_id: &str,
user_id: &str
) -> UserDeleteCall<'a, C, A>

Create a builder to help you perform the following task:

Deleted an EMM-managed user.

Arguments

enterpriseId - The ID of the enterprise.

userId - The ID of the user.

pub fn list(&self, enterprise_id: &str, email: &str) -> UserListCall<'a, C, A>

Create a builder to help you perform the following task:

Looks up a user by primary email address. This is only supported for Google-managed users. Lookup of the id is not needed for EMM-managed users because the id is already returned in the result of the Users.insert call.

Arguments

enterpriseId - The ID of the enterprise.

email - The exact primary email address of the user to look up.

pub fn generate_authentication_token(
&self,
enterprise_id: &str,
user_id: &str
) -> UserGenerateAuthenticationTokenCall<'a, C, A>

Create a builder to help you perform the following task:

Generates an authentication token which the device policy client can use to provision the given EMM-managed user account on a device. The generated token is single-use and expires after a few minutes.

You can provision a maximum of 10 devices per user.

This call only works with EMM-managed accounts.

Arguments

enterpriseId - The ID of the enterprise.

userId - The ID of the user.