Crate google_androidenterprise1

This documentation was generated from Android Enterprise crate version 0.1.8+20150309, where 20150309 is the exact revision of the androidenterprise:v1 schema built by the mako code generator v0.1.8. The original source code is on github.

Features

Handle the following Resources with ease from the central hub ...

collections

delete, get, insert, list, patch and update

collectionviewers

delete, get, list, patch and update

devices

get, get state, list and set state

enterprises

delete, enroll, get, insert, list, set account and unenroll

entitlements

delete, get, list, patch and update

grouplicenses

get and list

grouplicenseusers

list

installs

delete, get, list, patch and update

permissions

get

products

get, get app restrictions schema, get permissions and update permissions

users

generate token, get, list and revoke token

Not what you are looking for ? Find all other Google APIs in their Rust documentation index.

Structure of this Library

The API is structured into the following primary items:

Hub

a central object to maintain state and allow accessing all Activities

creates Method Builders which in turn allow access to individual Call Builders

Resources

primary types that you can apply Activities to

a collection of properties and Parts

Parts

a collection of properties

never directly used in Activities

Activities

operations to apply to Resources

All structures are marked with applicable traits to further categorize them and ease browsing.

Generally speaking, you can invoke Activities like this:

let r = hub.resource().activity(...).doit()

Or specifically ...

let r = hub.enterprises().insert(...).doit()
let r = hub.enterprises().list(...).doit()
let r = hub.enterprises().unenroll(...).doit()
let r = hub.enterprises().set_account(...).doit()
let r = hub.enterprises().enroll(...).doit()
let r = hub.enterprises().delete(...).doit()
let r = hub.enterprises().get(...).doit()

The resource() and activity(...) calls create builders. The second one dealing with Activities supports various methods to configure the impending operation (not shown here). It is made such that all required arguments have to be specified right away (i.e. (...)), whereas all optional ones can be build up as desired. The doit() method performs the actual communication with the server and returns the respective result.

Usage

Setting up your Project

To use this library, you would put the following lines into your Cargo.toml file:

[dependencies]
google-androidenterprise1 = "*"

A complete example

extern crate hyper;
extern crate yup_oauth2 as oauth2;
extern crate google_androidenterprise1 as androidenterprise1;
use androidenterprise1::Enterprise;
use androidenterprise1::{Result, Error};
use std::default::Default;
use oauth2::{Authenticator, DefaultAuthenticatorDelegate, ApplicationSecret, MemoryStorage};
use androidenterprise1::AndroidEnterprise;
 
// Get an ApplicationSecret instance by some means. It contains the `client_id` and 
// `client_secret`, among other things.
let secret: ApplicationSecret = Default::default();
// Instantiate the authenticator. It will choose a suitable authentication flow for you, 
// unless you replace  `None` with the desired Flow.
// Provide your own `AuthenticatorDelegate` to adjust the way it operates and get feedback about 
// what's going on. You probably want to bring in your own `TokenStorage` to persist tokens and
// retrieve them from storage.
let auth = Authenticator::new(&secret, DefaultAuthenticatorDelegate,
                              hyper::Client::new(),
                              <MemoryStorage as Default>::default(), None);
let mut hub = AndroidEnterprise::new(hyper::Client::new(), auth);
// As the method needs a request, you would usually fill it with the desired information
// into the respective structure. Some of the parts shown here might not be applicable !
// Values shown here are possibly random and not representative !
let mut req = Enterprise::default();
 
// You can configure optional parameters by calling the respective setters at will, and
// execute the final call using `doit()`.
// Values shown here are possibly random and not representative !
let result = hub.enterprises().insert(req, "token")
             .doit();
 
match result {
    Err(e) => match e {
        // The Error enum provides details about what exactly happened.
        // You can also just use its `Debug`, `Display` or `Error` traits
         Error::HttpError(_)
        |Error::MissingAPIKey
        |Error::MissingToken(_)
        |Error::Cancelled
        |Error::UploadSizeLimitExceeded(_, _)
        |Error::Failure(_)
        |Error::BadRequest(_)
        |Error::FieldClash(_)
        |Error::JsonDecodeError(_, _) => println!("{}", e),
    },
    Ok(res) => println!("Success: {:?}", res),
}

Handling Errors

All errors produced by the system are provided either as Result enumeration as return value of the doit() methods, or handed as possibly intermediate results to either the Hub Delegate, or the Authenticator Delegate.

When delegates handle errors or intermediate values, they may have a chance to instruct the system to retry. This makes the system potentially resilient to all kinds of errors.

Uploads and Downloads

If a method supports downloads, the response body, which is part of the Result, should be read by you to obtain the media. If such a method also supports a Response Result, it will return that by default. You can see it as meta-data for the actual media. To trigger a media download, you will have to set up the builder by making this call: .param("alt", "media").

Methods supporting uploads can do so using up to 2 different protocols: simple and resumable. The distinctiveness of each is represented by customized doit(...) methods, which are then named upload(...) and upload_resumable(...) respectively.

Customization and Callbacks

You may alter the way an doit() method is called by providing a delegate to the Method Builder before making the final doit() call. Respective methods will be called to provide progress information, as well as determine whether the system should retry on failure.

The delegate trait is default-implemented, allowing you to customize it with minimal effort.

Optional Parts in Server-Requests

All structures provided by this library are made to be enocodable and decodable via json. Optionals are used to indicate that partial requests are responses are valid. Most optionals are are considered Parts which are identifiable by name, which will be sent to the server to indicate either the set parts of the request or the desired parts in the response.

Builder Arguments

Using method builders, you are able to prepare an action call by repeatedly calling it's methods. These will always take a single argument, for which the following statements are true.

PODs are handed by copy

strings are passed as &str

request values are moved

Arguments will always be copied or cloned into the builder, to make them independent of their original life times.

Structs

AndroidEnterprise
Central instance to access all AndroidEnterprise related resource activities

AppRestrictionsSchema
Represents the list of app restrictions available to be pre-configured for the product.

AppRestrictionsSchemaRestriction
A restriction in the App Restriction Schema represents a piece of configuration that may be pre-applied.

AppRestrictionsSchemaRestrictionRestrictionValue
A typed value for the restriction.

Collection
A collection resource defines a named set of apps that is visible to a set of users in the Google Play Store app running on those users' managed devices. Those users can then install any of those apps if they wish (which will trigger creation of install and entitlement resources). A user cannot install an app on a managed device unless the app is listed in at least one collection that is visible to that user.

CollectionDeleteCall
Deletes a collection.

CollectionGetCall
Retrieves the details of a collection.

CollectionInsertCall
Creates a new collection.

CollectionListCall
Retrieves the IDs of all the collections for an enterprise.

CollectionMethods
A builder providing access to all methods supported on collection resources. It is not used directly, but through the AndroidEnterprise hub.

CollectionPatchCall
Updates a collection. This method supports patch semantics.

CollectionUpdateCall
Updates a collection.

CollectionViewersListResponse
The user resources for the collection.

CollectionsListResponse
The collection resources for the enterprise.

CollectionviewerDeleteCall
Removes the user from the list of those specifically allowed to see the collection. If the collection's visibility is set to viewersOnly then only such users will see the collection.

CollectionviewerGetCall
Retrieves the ID of the user if they have been specifically allowed to see the collection. If the collection's visibility is set to viewersOnly then only these users will see the collection.

CollectionviewerListCall
Retrieves the IDs of the users who have been specifically allowed to see the collection. If the collection's visibility is set to viewersOnly then only these users will see the collection.

CollectionviewerMethods
A builder providing access to all methods supported on collectionviewer resources. It is not used directly, but through the AndroidEnterprise hub.

CollectionviewerPatchCall
Adds the user to the list of those specifically allowed to see the collection. If the collection's visibility is set to viewersOnly then only such users will see the collection. This method supports patch semantics.

CollectionviewerUpdateCall
Adds the user to the list of those specifically allowed to see the collection. If the collection's visibility is set to viewersOnly then only such users will see the collection.

DefaultDelegate
A delegate with a conservative default implementation, which is used if no other delegate is set.

Device
A device resource represents a mobile device managed by the MDM and belonging to a specific enterprise user.

DeviceGetCall
Retrieves the details of a device.

DeviceGetStateCall
Retrieves whether a device is enabled or disabled for access by the user to Google services. The device state takes effect only if enforcing EMM policies on Android devices is enabled in the Google Admin Console. Otherwise, the device state is ignored and all devices are allowed access to Google services.

DeviceListCall
Retrieves the IDs of all of a user's devices.

DeviceMethods
A builder providing access to all methods supported on device resources. It is not used directly, but through the AndroidEnterprise hub.

DeviceSetStateCall
Sets whether a device is enabled or disabled for access by the user to Google services. The device state takes effect only if enforcing EMM policies on Android devices is enabled in the Google Admin Console. Otherwise, the device state is ignored and all devices are allowed access to Google services.

DeviceState
The state of a user's device, as accessed by the getState and setState methods on device resources.

DevicesListResponse
The device resources for the user.

Enterprise
An enterprise resource represents a binding between an organisation and their MDM.

EnterpriseAccount
A service account that can be used to authenticate as the enterprise to API calls that require such authentication.

EnterpriseDeleteCall
Deletes the binding between the MDM and enterprise. This is now deprecated; use this to unenroll customers that were previously enrolled with the 'insert' call, then enroll them again with the 'enroll' call.

EnterpriseEnrollCall
Enrolls an enterprise with the calling MDM.

EnterpriseGetCall
Retrieves the name and domain of an enterprise.

EnterpriseInsertCall
Establishes the binding between the MDM and an enterprise. This is now deprecated; use enroll instead.

EnterpriseListCall
Looks up an enterprise by domain name.

EnterpriseMethods
A builder providing access to all methods supported on enterprise resources. It is not used directly, but through the AndroidEnterprise hub.

EnterpriseSetAccountCall
Set the account that will be used to authenticate to the API as the enterprise.

EnterpriseUnenrollCall
Unenrolls an enterprise from the calling MDM.

EnterprisesListResponse
The matching enterprise resources.

Entitlement
The existence of an entitlement resource means that a user has the right to use a particular app on any of their devices. This might be because the app is free or because they have been allocated a license to the app from a group license purchased by the enterprise.

EntitlementDeleteCall
Removes an entitlement to an app for a user and uninstalls it.

EntitlementGetCall
Retrieves details of an entitlement.

EntitlementListCall
List of all entitlements for the specified user. Only the ID is set.

EntitlementMethods
A builder providing access to all methods supported on entitlement resources. It is not used directly, but through the AndroidEnterprise hub.

EntitlementPatchCall
Adds or updates an entitlement to an app for a user. This method supports patch semantics.

EntitlementUpdateCall
Adds or updates an entitlement to an app for a user.

EntitlementsListResponse
The entitlement resources for the user.

ErrorResponse
A utility to represent detailed errors we might see in case there are BadRequests. The latter happen if the sent parameters or request structures are unsound

GroupLicense
A group license object indicates a product that an enterprise admin has approved for use in the enterprise. The product may be free or paid. For free products, a group license object is created in these cases: if the enterprise admin approves a product in Google Play, if the product is added to a collection, or if an entitlement for the product is created for a user via the API. For paid products, a group license object is only created as part of the first bulk purchase of that product in Google Play done by the enterprise admin.

GroupLicenseUsersListResponse
The user resources for the group license.

GroupLicensesListResponse
The grouplicense resources for the enterprise.

GrouplicenseGetCall
Retrieves details of an enterprise's group license for a product.

GrouplicenseListCall
Retrieves IDs of all products for which the enterprise has a group license.

GrouplicenseMethods
A builder providing access to all methods supported on grouplicense resources. It is not used directly, but through the AndroidEnterprise hub.

GrouplicenseuserListCall
Retrieves the IDs of the users who have been granted entitlements under the license.

GrouplicenseuserMethods
A builder providing access to all methods supported on grouplicenseuser resources. It is not used directly, but through the AndroidEnterprise hub.

Install
The existence of an install resource indicates that an app is installed on a particular device (or that an install is pending).

InstallDeleteCall
Requests to remove an app from a device. A call to get or list will still show the app as installed on the device until it is actually removed.

InstallGetCall
Retrieves details of an installation of an app on a device.

InstallListCall
Retrieves the details of all apps installed on the specified device.

InstallMethods
A builder providing access to all methods supported on install resources. It is not used directly, but through the AndroidEnterprise hub.

InstallPatchCall
Requests to install the latest version of an app to a device. If the app is already installed then it is updated to the latest version if necessary. This method supports patch semantics.

InstallUpdateCall
Requests to install the latest version of an app to a device. If the app is already installed then it is updated to the latest version if necessary.

InstallsListResponse
The install resources for the device.

MethodInfo
Contains information about an API request.

MultiPartReader
Provides a Read interface that converts multiple parts into the protocol identified by RFC2387. Note: This implementation is just as rich as it needs to be to perform uploads to google APIs, and might not be a fully-featured implementation.

Permission
A permission represents some extra capability, to be granted to an Android app, which requires explicit consent. An enterprise admin must consent to these permissions on behalf of their users before an entitlement for the app can be created.

PermissionGetCall
Retrieves details of an Android app permission for display to an enterprise admin.

PermissionMethods
A builder providing access to all methods supported on permission resources. It is not used directly, but through the AndroidEnterprise hub.

Product
A product represents an app in the Google Play Store that is available to at least some users in the enterprise. (Some apps are restricted to a single enterprise, and no information about them is made available outside that enterprise.)

ProductGetAppRestrictionsSchemaCall
Retrieves the schema defining app restrictions configurable for this product. All products have a schema, but this may be empty if no app restrictions are defined.

ProductGetCall
Retrieves details of a product for display to an enterprise admin.

ProductGetPermissionCall
Retrieves the Android app permissions required by this app.

ProductMethods
A builder providing access to all methods supported on product resources. It is not used directly, but through the AndroidEnterprise hub.

ProductPermission
A product permissions resource represents the set of permissions required by a specific app and whether or not they have been accepted by an enterprise admin.

ProductPermissions
Information about the permissions required by a specific app and whether they have been accepted by the enterprise.

ProductUpdatePermissionCall
Updates the set of Android app permissions for this app that have been accepted by the enterprise.

User
A user resource represents an individual user within the enterprise's domain.

UserGenerateTokenCall
Generates a token (activation code) to allow this user to configure their work account in the Android Setup Wizard. Revokes any previously generated token.

UserGetCall
Retrieves a user's details.

UserListCall
Looks up a user by email address.

UserMethods
A builder providing access to all methods supported on user resources. It is not used directly, but through the AndroidEnterprise hub.

UserRevokeTokenCall
Revokes a previously generated token (activation code) for the user.

UserToken
A UserToken is used by a user when setting up a managed device or profile with their work account on a device. When the user enters their email address and token (activation code) the appropriate MDM app can be automatically downloaded.

UsersListResponse
The matching user resources.

Enums

Error

Scope
Identifies the an OAuth2 authorization scope. A scope is needed when requesting an authorization token.

Traits

CallBuilder
Identifies types which represent builders for a particular resource method

Delegate
A trait specifying functionality to help controlling any request performed by the API. The trait has a conservative default implementation.

Hub
Identifies the Hub. There is only one per library, this trait is supposed to make intended use more explicit. The hub allows to access all resource methods more easily.

MethodsBuilder
Identifies types for building methods of a particular resource type

NestedType
Identifies types which are only used by other types internally. They have no special meaning, this trait just marks them for completeness.

Part
Identifies types which are only used as part of other types, which usually are carrying the Resource trait.

ReadSeek
A utility to specify reader types which provide seeking capabilities too

RequestValue
Identifies types which are used in API requests.

Resource
Identifies types which can be inserted and deleted. Types with this trait are most commonly used by clients of this API.

ResponseResult
Identifies types which are used in API responses.

ToParts
A trait for all types that can convert themselves into a parts string

Functions

remove_json_null_values

Type Definitions

Result
A universal result type used as return for all calls.