Crate google_androidenterprise1

This documentation was generated from Android Enterprise crate version 1.0.9+20190624, where 20190624 is the exact revision of the androidenterprise:v1 schema built by the mako code generator v1.0.9.

Everything else about the Android Enterprise v1 API can be found at the official documentation site. The original source code is on github.

Features

Handle the following Resources with ease from the central hub ...

devices

force report upload, get, get state, list, patch, set state and update

enterprises

acknowledge notification set, complete signup, create web token, enroll, generate signup url, get, get service account, get store layout, list, pull notification set, send test push notification, set account, set store layout and unenroll

entitlements

delete, get, list, patch and update

grouplicenses

get and list

grouplicenseusers

list

installs

delete, get, list, patch and update

managedconfigurationsfordevice

delete, get, list, patch and update

managedconfigurationsforuser

delete, get, list, patch and update

managedconfigurationssettings

list

permissions

get

products

approve, generate approval url, get, get app restrictions schema, get permissions, list and unapprove

serviceaccountkeys

delete, insert and list

storelayoutclusters

delete, get, insert, list, patch and update

storelayoutpages

delete, get, insert, list, patch and update

users

delete, generate authentication token, generate token, get, get available product set, insert, list, patch, revoke device access, revoke token, set available product set and update

webapps

delete, get, insert, list, patch and update

Not what you are looking for ? Find all other Google APIs in their Rust documentation index.

Structure of this Library

The API is structured into the following primary items:

Hub

a central object to maintain state and allow accessing all Activities

creates Method Builders which in turn allow access to individual Call Builders

Resources

primary types that you can apply Activities to

a collection of properties and Parts

Parts

a collection of properties

never directly used in Activities

Activities

operations to apply to Resources

All structures are marked with applicable traits to further categorize them and ease browsing.

Generally speaking, you can invoke Activities like this:

let r = hub.resource().activity(...).doit()

Or specifically ...

let r = hub.enterprises().send_test_push_notification(...).doit()
let r = hub.enterprises().get_store_layout(...).doit()
let r = hub.enterprises().get_service_account(...).doit()
let r = hub.enterprises().acknowledge_notification_set(...).doit()
let r = hub.enterprises().set_store_layout(...).doit()
let r = hub.enterprises().complete_signup(...).doit()
let r = hub.enterprises().list(...).doit()
let r = hub.enterprises().unenroll(...).doit()
let r = hub.enterprises().set_account(...).doit()
let r = hub.enterprises().pull_notification_set(...).doit()
let r = hub.enterprises().enroll(...).doit()
let r = hub.enterprises().create_web_token(...).doit()
let r = hub.enterprises().get(...).doit()
let r = hub.enterprises().generate_signup_url(...).doit()

The resource() and activity(...) calls create builders. The second one dealing with Activities supports various methods to configure the impending operation (not shown here). It is made such that all required arguments have to be specified right away (i.e. (...)), whereas all optional ones can be build up as desired. The doit() method performs the actual communication with the server and returns the respective result.

Usage

Setting up your Project

To use this library, you would put the following lines into your Cargo.toml file:

[dependencies]
google-androidenterprise1 = "*"
# This project intentionally uses an old version of Hyper. See
# https://github.com/Byron/google-apis-rs/issues/173 for more
# information.
hyper = "^0.10"
hyper-rustls = "^0.6"
serde = "^1.0"
serde_json = "^1.0"
yup-oauth2 = "^1.0"

A complete example

extern crate hyper;
extern crate hyper_rustls;
extern crate yup_oauth2 as oauth2;
extern crate google_androidenterprise1 as androidenterprise1;
use androidenterprise1::{Result, Error};
use std::default::Default;
use oauth2::{Authenticator, DefaultAuthenticatorDelegate, ApplicationSecret, MemoryStorage};
use androidenterprise1::AndroidEnterprise;
 
// Get an ApplicationSecret instance by some means. It contains the `client_id` and 
// `client_secret`, among other things.
let secret: ApplicationSecret = Default::default();
// Instantiate the authenticator. It will choose a suitable authentication flow for you, 
// unless you replace  `None` with the desired Flow.
// Provide your own `AuthenticatorDelegate` to adjust the way it operates and get feedback about 
// what's going on. You probably want to bring in your own `TokenStorage` to persist tokens and
// retrieve them from storage.
let auth = Authenticator::new(&secret, DefaultAuthenticatorDelegate,
                              hyper::Client::with_connector(hyper::net::HttpsConnector::new(hyper_rustls::TlsClient::new())),
                              <MemoryStorage as Default>::default(), None);
let mut hub = AndroidEnterprise::new(hyper::Client::with_connector(hyper::net::HttpsConnector::new(hyper_rustls::TlsClient::new())), auth);
// You can configure optional parameters by calling the respective setters at will, and
// execute the final call using `doit()`.
// Values shown here are possibly random and not representative !
let result = hub.enterprises().get_service_account("enterpriseId")
             .key_type("sed")
             .doit();
 
match result {
    Err(e) => match e {
        // The Error enum provides details about what exactly happened.
        // You can also just use its `Debug`, `Display` or `Error` traits
         Error::HttpError(_)
        |Error::MissingAPIKey
        |Error::MissingToken(_)
        |Error::Cancelled
        |Error::UploadSizeLimitExceeded(_, _)
        |Error::Failure(_)
        |Error::BadRequest(_)
        |Error::FieldClash(_)
        |Error::JsonDecodeError(_, _) => println!("{}", e),
    },
    Ok(res) => println!("Success: {:?}", res),
}

Handling Errors

All errors produced by the system are provided either as Result enumeration as return value of the doit() methods, or handed as possibly intermediate results to either the Hub Delegate, or the Authenticator Delegate.

When delegates handle errors or intermediate values, they may have a chance to instruct the system to retry. This makes the system potentially resilient to all kinds of errors.

Uploads and Downloads

If a method supports downloads, the response body, which is part of the Result, should be read by you to obtain the media. If such a method also supports a Response Result, it will return that by default. You can see it as meta-data for the actual media. To trigger a media download, you will have to set up the builder by making this call: .param("alt", "media").

Methods supporting uploads can do so using up to 2 different protocols: simple and resumable. The distinctiveness of each is represented by customized doit(...) methods, which are then named upload(...) and upload_resumable(...) respectively.

Customization and Callbacks

You may alter the way an doit() method is called by providing a delegate to the Method Builder before making the final doit() call. Respective methods will be called to provide progress information, as well as determine whether the system should retry on failure.

The delegate trait is default-implemented, allowing you to customize it with minimal effort.

Optional Parts in Server-Requests

All structures provided by this library are made to be enocodable and decodable via json. Optionals are used to indicate that partial requests are responses are valid. Most optionals are are considered Parts which are identifiable by name, which will be sent to the server to indicate either the set parts of the request or the desired parts in the response.

Builder Arguments

Using method builders, you are able to prepare an action call by repeatedly calling it's methods. These will always take a single argument, for which the following statements are true.

PODs are handed by copy

strings are passed as &str

request values are moved

Arguments will always be copied or cloned into the builder, to make them independent of their original life times.

Structs

Administrator
This represents an enterprise admin who can manage the enterprise in the managed Google Play store.

AdministratorWebToken
A token authorizing an admin to access an iframe.

AdministratorWebTokenSpec
Specification for a token used to generate iframes. The token specifies what data the admin is allowed to modify and the URI the iframe is allowed to communiate with.

AdministratorWebTokenSpecManagedConfigurations
There is no detailed description.

AdministratorWebTokenSpecPlaySearch
There is no detailed description.

AdministratorWebTokenSpecPrivateApps
There is no detailed description.

AdministratorWebTokenSpecStoreBuilder
There is no detailed description.

AdministratorWebTokenSpecWebApps
There is no detailed description.

AndroidEnterprise
Central instance to access all AndroidEnterprise related resource activities

AppRestrictionsSchema
Represents the list of app restrictions available to be pre-configured for the product.

AppRestrictionsSchemaChangeEvent
An event generated when a new app version is uploaded to Google Play and its app restrictions schema changed. To fetch the app restrictions schema for an app, use Products.getAppRestrictionsSchema on the EMM API.

AppRestrictionsSchemaRestriction
A restriction in the App Restriction Schema represents a piece of configuration that may be pre-applied.

AppRestrictionsSchemaRestrictionRestrictionValue
A typed value for the restriction.

AppState
List of states set by the app.

AppUpdateEvent
An event generated when a new version of an app is uploaded to Google Play. Notifications are sent for new public versions only: alpha, beta, or canary versions do not generate this event. To fetch up-to-date version history for an app, use Products.Get on the EMM API.

AppVersion
This represents a single version of the app.

ApprovalUrlInfo
Information on an approval URL.

AuthenticationToken
An AuthenticationToken is used by the EMM's device policy client on a device to provision the given EMM-managed user on that device.

AutoInstallConstraint
The auto-install constraint. Defines a set of restrictions for installation. At least one of the fields must be set.

AutoInstallPolicy
There is no detailed description.

Chunk

ConfigurationVariables
A configuration variables resource contains the managed configuration settings ID to be applied to a single user, as well as the variable set that is attributed to the user. The variable set will be used to replace placeholders in the managed configuration settings.

ContentRange
Implements the Content-Range header, for serialization only

DefaultDelegate
A delegate with a conservative default implementation, which is used if no other delegate is set.

Device
A Devices resource represents a mobile device managed by the EMM and belonging to a specific enterprise user.

DeviceForceReportUploadCall
Uploads a report containing any changes in app states on the device since the last report was generated. You can call this method up to 3 times every 24 hours for a given device.

DeviceGetCall
Retrieves the details of a device.

DeviceGetStateCall
Retrieves whether a device's access to Google services is enabled or disabled. The device state takes effect only if enforcing EMM policies on Android devices is enabled in the Google Admin Console. Otherwise, the device state is ignored and all devices are allowed access to Google services. This is only supported for Google-managed users.

DeviceListCall
Retrieves the IDs of all of a user's devices.

DeviceMethods
A builder providing access to all methods supported on device resources. It is not used directly, but through the AndroidEnterprise hub.

DevicePatchCall
Updates the device policy. This method supports patch semantics.

DeviceReport
Device report updated with the latest app states for managed apps on the device.

DeviceReportUpdateEvent
An event generated when an updated device report is available.

DeviceSetStateCall
Sets whether a device's access to Google services is enabled or disabled. The device state takes effect only if enforcing EMM policies on Android devices is enabled in the Google Admin Console. Otherwise, the device state is ignored and all devices are allowed access to Google services. This is only supported for Google-managed users.

DeviceState
The state of a user's device, as accessed by the getState and setState methods on device resources.

DeviceUpdateCall
Updates the device policy

DevicesListResponse
The device resources for the user.

DummyNetworkStream

Enterprise
An Enterprises resource represents the binding between an EMM and a specific organization. That binding can be instantiated in one of two different ways using this API as follows:

EnterpriseAccount
A service account that can be used to authenticate as the enterprise to API calls that require such authentication.

EnterpriseAcknowledgeNotificationSetCall
Acknowledges notifications that were received from Enterprises.PullNotificationSet to prevent subsequent calls from returning the same notifications.

EnterpriseCompleteSignupCall
Completes the signup flow, by specifying the Completion token and Enterprise token. This request must not be called multiple times for a given Enterprise Token.

EnterpriseCreateWebTokenCall
Returns a unique token to access an embeddable UI. To generate a web UI, pass the generated token into the managed Google Play javascript API. Each token may only be used to start one UI session. See the javascript API documentation for further information.

EnterpriseEnrollCall
Enrolls an enterprise with the calling EMM.

EnterpriseGenerateSignupUrlCall
Generates a sign-up URL.

EnterpriseGetCall
Retrieves the name and domain of an enterprise.

EnterpriseGetServiceAccountCall
Returns a service account and credentials. The service account can be bound to the enterprise by calling setAccount. The service account is unique to this enterprise and EMM, and will be deleted if the enterprise is unbound. The credentials contain private key data and are not stored server-side.

EnterpriseGetStoreLayoutCall
Returns the store layout for the enterprise. If the store layout has not been set, returns "basic" as the store layout type and no homepage.

EnterpriseListCall
Looks up an enterprise by domain name. This is only supported for enterprises created via the Google-initiated creation flow. Lookup of the id is not needed for enterprises created via the EMM-initiated flow since the EMM learns the enterprise ID in the callback specified in the Enterprises.generateSignupUrl call.

EnterpriseMethods
A builder providing access to all methods supported on enterprise resources. It is not used directly, but through the AndroidEnterprise hub.

EnterprisePullNotificationSetCall
Pulls and returns a notification set for the enterprises associated with the service account authenticated for the request. The notification set may be empty if no notification are pending. A notification set returned needs to be acknowledged within 20 seconds by calling Enterprises.AcknowledgeNotificationSet, unless the notification set is empty. Notifications that are not acknowledged within the 20 seconds will eventually be included again in the response to another PullNotificationSet request, and those that are never acknowledged will ultimately be deleted according to the Google Cloud Platform Pub/Sub system policy. Multiple requests might be performed concurrently to retrieve notifications, in which case the pending notifications (if any) will be split among each caller, if any are pending. If no notifications are present, an empty notification list is returned. Subsequent requests may return more notifications once they become available.

EnterpriseSendTestPushNotificationCall
Sends a test notification to validate the EMM integration with the Google Cloud Pub/Sub service for this enterprise.

EnterpriseSetAccountCall
Sets the account that will be used to authenticate to the API as the enterprise.

EnterpriseSetStoreLayoutCall
Sets the store layout for the enterprise. By default, storeLayoutType is set to "basic" and the basic store layout is enabled. The basic layout only contains apps approved by the admin, and that have been added to the available product set for a user (using the setAvailableProductSet call). Apps on the page are sorted in order of their product ID value. If you create a custom store layout (by setting storeLayoutType = "custom" and setting a homepage), the basic store layout is disabled.

EnterpriseUnenrollCall
Unenrolls an enterprise from the calling EMM.

EnterprisesListResponse
The matching enterprise resources.

EnterprisesSendTestPushNotificationResponse
There is no detailed description.

Entitlement
The presence of an Entitlements resource indicates that a user has the right to use a particular app. Entitlements are user specific, not device specific. This allows a user with an entitlement to an app to install the app on all their devices. It's also possible for a user to hold an entitlement to an app without installing the app on any device.

EntitlementDeleteCall
Removes an entitlement to an app for a user.

EntitlementGetCall
Retrieves details of an entitlement.

EntitlementListCall
Lists all entitlements for the specified user. Only the ID is set.

EntitlementMethods
A builder providing access to all methods supported on entitlement resources. It is not used directly, but through the AndroidEnterprise hub.

EntitlementPatchCall
Adds or updates an entitlement to an app for a user. This method supports patch semantics.

EntitlementUpdateCall
Adds or updates an entitlement to an app for a user.

EntitlementsListResponse
The entitlement resources for the user.

ErrorResponse
A utility to represent detailed errors we might see in case there are BadRequests. The latter happen if the sent parameters or request structures are unsound

GroupLicense
Group license objects allow you to keep track of licenses (called entitlements) for both free and paid apps. For a free app, a group license is created when an enterprise admin first approves the product in Google Play or when the first entitlement for the product is created for a user via the API. For a paid app, a group license object is only created when an enterprise admin purchases the product in Google Play for the first time.

GroupLicenseUsersListResponse
The user resources for the group license.

GroupLicensesListResponse
The grouplicense resources for the enterprise.

GrouplicenseGetCall
Retrieves details of an enterprise's group license for a product.

GrouplicenseListCall
Retrieves IDs of all products for which the enterprise has a group license.

GrouplicenseMethods
A builder providing access to all methods supported on grouplicense resources. It is not used directly, but through the AndroidEnterprise hub.

GrouplicenseuserListCall
Retrieves the IDs of the users who have been granted entitlements under the license.

GrouplicenseuserMethods
A builder providing access to all methods supported on grouplicenseuser resources. It is not used directly, but through the AndroidEnterprise hub.

Install
The existence of an Installs resource indicates that an app is installed on a particular device (or that an install is pending).

InstallDeleteCall
Requests to remove an app from a device. A call to get or list will still show the app as installed on the device until it is actually removed.

InstallFailureEvent
An event generated when an app installation failed on a device

InstallGetCall
Retrieves details of an installation of an app on a device.

InstallListCall
Retrieves the details of all apps installed on the specified device.

InstallMethods
A builder providing access to all methods supported on install resources. It is not used directly, but through the AndroidEnterprise hub.

InstallPatchCall
Requests to install the latest version of an app to a device. If the app is already installed, then it is updated to the latest version if necessary. This method supports patch semantics.

InstallUpdateCall
Requests to install the latest version of an app to a device. If the app is already installed, then it is updated to the latest version if necessary.

InstallsListResponse
The install resources for the device.

JsonServerError
A utility type which can decode a server response that indicates error

KeyedAppState
Represents a keyed app state containing a key, timestamp, severity level, optional description, and optional data.

LocalizedText
A localized string with its locale.

MaintenanceWindow
Maintenance window for managed Google Play Accounts. This allows Play store to update the apps on the foreground in the designated window.

ManagedConfiguration
A managed configuration resource contains the set of managed properties defined by the app developer in the app's managed configurations schema, as well as any configuration variables defined for the user.

ManagedConfigurationsForDeviceListResponse
The managed configuration resources for the device.

ManagedConfigurationsForUserListResponse
The managed configuration resources for the user.

ManagedConfigurationsSettings
A managed configurations settings resource contains the set of managed properties that have been configured for an Android app to be applied to a set of users. The app's developer would have defined configurable properties in the managed configurations schema.

ManagedConfigurationsSettingsListResponse
The managed configurations settings for a product.

ManagedProperty
A managed property of a managed configuration. The property must match one of the properties in the app restrictions schema of the product. Exactly one of the value fields must be populated, and it must match the property's type in the app restrictions schema.

ManagedPropertyBundle
A bundle of managed properties.

ManagedconfigurationsfordeviceDeleteCall
Removes a per-device managed configuration for an app for the specified device.

ManagedconfigurationsfordeviceGetCall
Retrieves details of a per-device managed configuration.

ManagedconfigurationsfordeviceListCall
Lists all the per-device managed configurations for the specified device. Only the ID is set.

ManagedconfigurationsfordeviceMethods
A builder providing access to all methods supported on managedconfigurationsfordevice resources. It is not used directly, but through the AndroidEnterprise hub.

ManagedconfigurationsfordevicePatchCall
Adds or updates a per-device managed configuration for an app for the specified device. This method supports patch semantics.

ManagedconfigurationsfordeviceUpdateCall
Adds or updates a per-device managed configuration for an app for the specified device.

ManagedconfigurationsforuserDeleteCall
Removes a per-user managed configuration for an app for the specified user.

ManagedconfigurationsforuserGetCall
Retrieves details of a per-user managed configuration for an app for the specified user.

ManagedconfigurationsforuserListCall
Lists all the per-user managed configurations for the specified user. Only the ID is set.

ManagedconfigurationsforuserMethods
A builder providing access to all methods supported on managedconfigurationsforuser resources. It is not used directly, but through the AndroidEnterprise hub.

ManagedconfigurationsforuserPatchCall
Adds or updates the managed configuration settings for an app for the specified user. If you support the Managed configurations iframe, you can apply managed configurations to a user by specifying an mcmId and its associated configuration variables (if any) in the request. Alternatively, all EMMs can apply managed configurations by passing a list of managed properties. This method supports patch semantics.

ManagedconfigurationsforuserUpdateCall
Adds or updates the managed configuration settings for an app for the specified user. If you support the Managed configurations iframe, you can apply managed configurations to a user by specifying an mcmId and its associated configuration variables (if any) in the request. Alternatively, all EMMs can apply managed configurations by passing a list of managed properties.

ManagedconfigurationssettingListCall
Lists all the managed configurations settings for the specified app. Only the ID and the name is set.

ManagedconfigurationssettingMethods
A builder providing access to all methods supported on managedconfigurationssetting resources. It is not used directly, but through the AndroidEnterprise hub.

MethodInfo
Contains information about an API request.

MultiPartReader
Provides a Read interface that converts multiple parts into the protocol identified by RFC2387. Note: This implementation is just as rich as it needs to be to perform uploads to google APIs, and might not be a fully-featured implementation.

NewDeviceEvent
An event generated when a new device is ready to be managed.

NewPermissionsEvent
An event generated when new permissions are added to an app.

Notification
A notification of one event relating to an enterprise.

NotificationSet
A resource returned by the PullNotificationSet API, which contains a collection of notifications for enterprises associated with the service account authenticated for the request.

PageInfo
There is no detailed description.

Permission
A Permissions resource represents some extra capability, to be granted to an Android app, which requires explicit consent. An enterprise admin must consent to these permissions on behalf of their users before an entitlement for the app can be created.

PermissionGetCall
Retrieves details of an Android app permission for display to an enterprise admin.

PermissionMethods
A builder providing access to all methods supported on permission resources. It is not used directly, but through the AndroidEnterprise hub.

Policy
The device policy for a given managed device.

Product
A Products resource represents an app in the Google Play store that is available to at least some users in the enterprise. (Some apps are restricted to a single enterprise, and no information about them is made available outside that enterprise.)

ProductApprovalEvent
An event generated when a product's approval status is changed.

ProductApproveCall
Approves the specified product and the relevant app permissions, if any. The maximum number of products that you can approve per enterprise customer is 1,000.

ProductAvailabilityChangeEvent
An event generated whenever a product's availability changes.

ProductGenerateApprovalUrlCall
Generates a URL that can be rendered in an iframe to display the permissions (if any) of a product. An enterprise admin must view these permissions and accept them on behalf of their organization in order to approve that product.

ProductGetAppRestrictionsSchemaCall
Retrieves the schema that defines the configurable properties for this product. All products have a schema, but this schema may be empty if no managed configurations have been defined. This schema can be used to populate a UI that allows an admin to configure the product. To apply a managed configuration based on the schema obtained using this API, see Managed Configurations through Play.

ProductGetCall
Retrieves details of a product for display to an enterprise admin.

ProductGetPermissionCall
Retrieves the Android app permissions required by this app.

ProductListCall
Finds approved products that match a query, or all approved products if there is no query.

ProductMethods
A builder providing access to all methods supported on product resources. It is not used directly, but through the AndroidEnterprise hub.

ProductPermission
A product permissions resource represents the set of permissions required by a specific app and whether or not they have been accepted by an enterprise admin.

ProductPermissions
Information about the permissions required by a specific app and whether they have been accepted by the enterprise.

ProductPolicy
The policy for a product.

ProductSet
A set of products.

ProductSigningCertificate
There is no detailed description.

ProductUnapproveCall
Unapproves the specified product (and the relevant app permissions, if any)

ProductVisibility
A product to be made visible to a user.

ProductsApproveRequest
There is no detailed description.

ProductsGenerateApprovalUrlResponse
There is no detailed description.

ProductsListResponse
The matching products.

RangeResponseHeader

ResumableUploadHelper
A utility type to perform a resumable upload from start to end.

ServerError

ServerMessage

ServiceAccount
A service account identity, including the name and credentials that can be used to authenticate as the service account.

ServiceAccountKey
Credentials that can be used to authenticate as a service account.

ServiceAccountKeysListResponse
There is no detailed description.

ServiceaccountkeyDeleteCall
Removes and invalidates the specified credentials for the service account associated with this enterprise. The calling service account must have been retrieved by calling Enterprises.GetServiceAccount and must have been set as the enterprise service account by calling Enterprises.SetAccount.

ServiceaccountkeyInsertCall
Generates new credentials for the service account associated with this enterprise. The calling service account must have been retrieved by calling Enterprises.GetServiceAccount and must have been set as the enterprise service account by calling Enterprises.SetAccount.

ServiceaccountkeyListCall
Lists all active credentials for the service account associated with this enterprise. Only the ID and key type are returned. The calling service account must have been retrieved by calling Enterprises.GetServiceAccount and must have been set as the enterprise service account by calling Enterprises.SetAccount.

ServiceaccountkeyMethods
A builder providing access to all methods supported on serviceaccountkey resources. It is not used directly, but through the AndroidEnterprise hub.

SignupInfo
A resource returned by the GenerateSignupUrl API, which contains the Signup URL and Completion Token.

StoreCluster
Definition of a managed Google Play store cluster, a list of products displayed as part of a store page.

StoreLayout
General setting for the managed Google Play store layout, currently only specifying the page to display the first time the store is opened.

StoreLayoutClustersListResponse
The store page resources for the enterprise.

StoreLayoutPagesListResponse
The store page resources for the enterprise.

StorePage
Definition of a managed Google Play store page, made of a localized name and links to other pages. A page also contains clusters defined as a subcollection.

StorelayoutclusterDeleteCall
Deletes a cluster.

StorelayoutclusterGetCall
Retrieves details of a cluster.

StorelayoutclusterInsertCall
Inserts a new cluster in a page.

StorelayoutclusterListCall
Retrieves the details of all clusters on the specified page.

StorelayoutclusterMethods
A builder providing access to all methods supported on storelayoutcluster resources. It is not used directly, but through the AndroidEnterprise hub.

StorelayoutclusterPatchCall
Updates a cluster. This method supports patch semantics.

StorelayoutclusterUpdateCall
Updates a cluster.

StorelayoutpageDeleteCall
Deletes a store page.

StorelayoutpageGetCall
Retrieves details of a store page.

StorelayoutpageInsertCall
Inserts a new store page.

StorelayoutpageListCall
Retrieves the details of all pages in the store.

StorelayoutpageMethods
A builder providing access to all methods supported on storelayoutpage resources. It is not used directly, but through the AndroidEnterprise hub.

StorelayoutpagePatchCall
Updates the content of a store page. This method supports patch semantics.

StorelayoutpageUpdateCall
Updates the content of a store page.

TokenPagination
There is no detailed description.

TrackInfo
Id to name association of a track.

User
A Users resource represents an account associated with an enterprise. The account may be specific to a device or to an individual user (who can then use the account across multiple devices). The account may provide access to managed Google Play only, or to other Google services, depending on the identity model:

UserDeleteCall
Deleted an EMM-managed user.

UserGenerateAuthenticationTokenCall
Generates an authentication token which the device policy client can use to provision the given EMM-managed user account on a device. The generated token is single-use and expires after a few minutes.

UserGenerateTokenCall
Generates a token (activation code) to allow this user to configure their managed account in the Android Setup Wizard. Revokes any previously generated token.

UserGetAvailableProductSetCall
Retrieves the set of products a user is entitled to access.

UserGetCall
Retrieves a user's details.

UserInsertCall
Creates a new EMM-managed user.

UserListCall
Looks up a user by primary email address. This is only supported for Google-managed users. Lookup of the id is not needed for EMM-managed users because the id is already returned in the result of the Users.insert call.

UserMethods
A builder providing access to all methods supported on user resources. It is not used directly, but through the AndroidEnterprise hub.

UserPatchCall
Updates the details of an EMM-managed user.

UserRevokeDeviceAccesCall
Revokes access to all devices currently provisioned to the user. The user will no longer be able to use the managed Play store on any of their managed devices.

UserRevokeTokenCall
Revokes a previously generated token (activation code) for the user.

UserSetAvailableProductSetCall
Modifies the set of products that a user is entitled to access (referred to as whitelisted products). Only products that are approved or products that were previously approved (products with revoked approval) can be whitelisted.

UserToken
A UserToken is used by a user when setting up a managed device or profile with their managed Google Play account on a device. When the user enters their email address and token (activation code) the appropriate EMM app can be automatically downloaded.

UserUpdateCall
Updates the details of an EMM-managed user.

UsersListResponse
The matching user resources.

VariableSet
A variable set is a key-value pair of EMM-provided placeholders and its corresponding value, which is attributed to a user. For example, $FIRSTNAME could be a placeholder, and its value could be Alice. Placeholders should start with a '$' sign and should be alphanumeric only.

WebApp
A WebApps resource represents a web app created for an enterprise. Web apps are published to managed Google Play and can be distributed like other Android apps. On a user's device, a web app opens its specified URL.

WebAppIcon
Icon for a web app.

WebAppsListResponse
The web app details for an enterprise.

WebappDeleteCall
Deletes an existing web app.

WebappGetCall
Gets an existing web app.

WebappInsertCall
Creates a new web app for the enterprise.

WebappListCall
Retrieves the details of all web apps for a given enterprise.

WebappMethods
A builder providing access to all methods supported on webapp resources. It is not used directly, but through the AndroidEnterprise hub.

WebappPatchCall
Updates an existing web app. This method supports patch semantics.

WebappUpdateCall
Updates an existing web app.

XUploadContentType
The X-Upload-Content-Type header.

Enums

Error

Scope
Identifies the an OAuth2 authorization scope. A scope is needed when requesting an authorization token.

Traits

CallBuilder
Identifies types which represent builders for a particular resource method

Delegate
A trait specifying functionality to help controlling any request performed by the API. The trait has a conservative default implementation.

Hub
Identifies the Hub. There is only one per library, this trait is supposed to make intended use more explicit. The hub allows to access all resource methods more easily.

MethodsBuilder
Identifies types for building methods of a particular resource type

NestedType
Identifies types which are only used by other types internally. They have no special meaning, this trait just marks them for completeness.

Part
Identifies types which are only used as part of other types, which usually are carrying the Resource trait.

ReadSeek
A utility to specify reader types which provide seeking capabilities too

RequestValue
Identifies types which are used in API requests.

Resource
Identifies types which can be inserted and deleted. Types with this trait are most commonly used by clients of this API.

ResponseResult
Identifies types which are used in API responses.

ToParts
A trait for all types that can convert themselves into a parts string

UnusedType
Identifies types which are not actually used by the API This might be a bug within the google API schema.

Functions

remove_json_null_values

Type Definitions

Result
A universal result type used as return for all calls.