Struct google_androidenterprise1::AndroidEnterprise

pub struct AndroidEnterprise<C, A> {
    // some fields omitted
}

Central instance to access all AndroidEnterprise related resource activities

Examples

Instantiate a new hub

extern crate hyper;
extern crate yup_oauth2 as oauth2;
extern crate google_androidenterprise1 as androidenterprise1;
use androidenterprise1::StoreLayout;
use androidenterprise1::{Result, Error};
use std::default::Default;
use oauth2::{Authenticator, DefaultAuthenticatorDelegate, ApplicationSecret, MemoryStorage};
use androidenterprise1::AndroidEnterprise;
 
// Get an ApplicationSecret instance by some means. It contains the `client_id` and 
// `client_secret`, among other things.
let secret: ApplicationSecret = Default::default();
// Instantiate the authenticator. It will choose a suitable authentication flow for you, 
// unless you replace  `None` with the desired Flow.
// Provide your own `AuthenticatorDelegate` to adjust the way it operates and get feedback about 
// what's going on. You probably want to bring in your own `TokenStorage` to persist tokens and
// retrieve them from storage.
let auth = Authenticator::new(&secret, DefaultAuthenticatorDelegate,
                              hyper::Client::new(),
                              <MemoryStorage as Default>::default(), None);
let mut hub = AndroidEnterprise::new(hyper::Client::new(), auth);
// As the method needs a request, you would usually fill it with the desired information
// into the respective structure. Some of the parts shown here might not be applicable !
// Values shown here are possibly random and not representative !
let mut req = StoreLayout::default();
 
// You can configure optional parameters by calling the respective setters at will, and
// execute the final call using `doit()`.
// Values shown here are possibly random and not representative !
let result = hub.enterprises().set_store_layout(req, "enterpriseId")
             .doit();
 
match result {
    Err(e) => match e {
        // The Error enum provides details about what exactly happened.
        // You can also just use its `Debug`, `Display` or `Error` traits
         Error::HttpError(_)
        |Error::MissingAPIKey
        |Error::MissingToken(_)
        |Error::Cancelled
        |Error::UploadSizeLimitExceeded(_, _)
        |Error::Failure(_)
        |Error::BadRequest(_)
        |Error::FieldClash(_)
        |Error::JsonDecodeError(_, _) => println!("{}", e),
    },
    Ok(res) => println!("Success: {:?}", res),
}

Methods

impl<'a, C, A> AndroidEnterprise<C, A> where C: BorrowMut<Client>, A: GetToken

fn new(client: C, authenticator: A) -> AndroidEnterprise<C, A>

fn collections(&'a self) -> CollectionMethods<'a, C, A>

fn collectionviewers(&'a self) -> CollectionviewerMethods<'a, C, A>

fn devices(&'a self) -> DeviceMethods<'a, C, A>

fn enterprises(&'a self) -> EnterpriseMethods<'a, C, A>

fn entitlements(&'a self) -> EntitlementMethods<'a, C, A>

fn grouplicenses(&'a self) -> GrouplicenseMethods<'a, C, A>

fn grouplicenseusers(&'a self) -> GrouplicenseuserMethods<'a, C, A>

fn installs(&'a self) -> InstallMethods<'a, C, A>

fn permissions(&'a self) -> PermissionMethods<'a, C, A>

fn products(&'a self) -> ProductMethods<'a, C, A>

fn storelayoutclusters(&'a self) -> StorelayoutclusterMethods<'a, C, A>

fn storelayoutpages(&'a self) -> StorelayoutpageMethods<'a, C, A>

fn users(&'a self) -> UserMethods<'a, C, A>

fn user_agent(&mut self, agent_name: String) -> String

Set the user-agent header field to use in all requests to the server. It defaults to google-api-rust-client/0.1.14.

Returns the previously set user-agent.