Struct google_androidenterprise1::Product

pub struct Product {
    pub small_icon_url: Option<String>,
    pub kind: Option<String>,
    pub distribution_channel: Option<String>,
    pub title: Option<String>,
    pub work_details_url: Option<String>,
    pub app_version: Option<Vec<AppVersion>>,
    pub product_pricing: Option<String>,
    pub icon_url: Option<String>,
    pub author_name: Option<String>,
    pub requires_container_app: Option<bool>,
    pub details_url: Option<String>,
    pub product_id: Option<String>,
}

A product represents an app in the Google Play Store that is available to at least some users in the enterprise. (Some apps are restricted to a single enterprise, and no information about them is made available outside that enterprise.)

The information provided for each product (localized name, icon, link to the full Google Play details page) is intended to allow a basic representation of the product within an EMM user interface.

Activities

This type is used in activities, which are methods you may call on this type or where this type is involved in. The list links the activity name, along with information about where it is used (one of request and response).

get permissions products (none)

get products (response)

approve products (none)

get app restrictions schema products (none)

generate approval url products (none)

update permissions products (none)

Fields

small_icon_url

A link to a smaller image that can be used as an icon for the product. This image is suitable for use at up to 128px x 128px.

kind

Identifies what kind of resource this is. Value: the fixed string "androidenterprise#product".

distribution_channel

How and to whom the package is made available. The value publicGoogleHosted means that the package is available through the Play Store and not restricted to a specific enterprise. The value privateGoogleHosted means that the package is a private app (restricted to an enterprise) but hosted by Google. The value privateSelfHosted means that the package is a private app (restricted to an enterprise) and is privately hosted.

title

The name of the product.

work_details_url

A link to the Google Play for Work details page for the product, for use by an Enterprise administrator.

app_version

App versions currently available for this product. The returned list contains only public versions. Alpha and beta versions are not included.

product_pricing

Whether this product is free, free with in-app purchases, or paid.

icon_url

A link to an image that can be used as an icon for the product. This image is suitable for use at up to 512px x 512px.

author_name

The name of the author of the product (e.g. the app developer).

requires_container_app

Whether this app can only be installed on devices using the Android for Work container app.

details_url

A link to the (consumer) Google Play details page for the product.

product_id

A string of the form app:. For example, app:com.google.android.gm represents the Gmail app.