Struct google_androidenterprise1::UserMethods

pub struct UserMethods<'a, C, A> where C: 'a, A: 'a {
    // some fields omitted
}

A builder providing access to all methods supported on user resources. It is not used directly, but through the AndroidEnterprise hub.

Example

Instantiate a resource builder

extern crate hyper;
extern crate yup_oauth2 as oauth2;
extern crate google_androidenterprise1 as androidenterprise1;
 
use std::default::Default;
use oauth2::{Authenticator, DefaultAuthenticatorDelegate, ApplicationSecret, MemoryStorage};
use androidenterprise1::AndroidEnterprise;
 
let secret: ApplicationSecret = Default::default();
let auth = Authenticator::new(&secret, DefaultAuthenticatorDelegate,
                              hyper::Client::new(),
                              <MemoryStorage as Default>::default(), None);
let mut hub = AndroidEnterprise::new(hyper::Client::new(), auth);
// Usually you wouldn't bind this to a variable, but keep calling *CallBuilders*
// like `generate_token(...)`, `get(...)`, `get_available_product_set(...)`, `list(...)`, `revoke_token(...)` and `set_available_product_set(...)`
// to build up your call.
let rb = hub.users();

Methods

impl<'a, C, A> UserMethods<'a, C, A>

fn revoke_token(&self, enterprise_id: &str, user_id: &str) -> UserRevokeTokenCall<'a, C, A>

Create a builder to help you perform the following task:

Revokes a previously generated token (activation code) for the user.

Arguments

enterpriseId - The ID of the enterprise.

userId - The ID of the user.

fn get(&self, enterprise_id: &str, user_id: &str) -> UserGetCall<'a, C, A>

Create a builder to help you perform the following task:

Retrieves a user's details.

Arguments

enterpriseId - The ID of the enterprise.

userId - The ID of the user.

fn list(&self, enterprise_id: &str, email: &str) -> UserListCall<'a, C, A>

Create a builder to help you perform the following task:

Looks up a user by email address. This only works for Google managed users.

Arguments

enterpriseId - The ID of the enterprise.

email - The exact primary email address of the user to look up.

fn generate_token(&self, enterprise_id: &str, user_id: &str) -> UserGenerateTokenCall<'a, C, A>

Create a builder to help you perform the following task:

Generates a token (activation code) to allow this user to configure their work account in the Android Setup Wizard. Revokes any previously generated token.

Arguments

enterpriseId - The ID of the enterprise.

userId - The ID of the user.

fn set_available_product_set(&self, request: ProductSet, enterprise_id: &str, user_id: &str) -> UserSetAvailableProductSetCall<'a, C, A>

Create a builder to help you perform the following task:

Modifies the set of products a user is entitled to access.

Arguments

request - No description provided.

enterpriseId - The ID of the enterprise.

userId - The ID of the user.

fn get_available_product_set(&self, enterprise_id: &str, user_id: &str) -> UserGetAvailableProductSetCall<'a, C, A>

Create a builder to help you perform the following task:

Retrieves the set of products a user is entitled to access.

Arguments

enterpriseId - The ID of the enterprise.

userId - The ID of the user.